requirements.txt
Django==1.7.7
Django-Eventbrite==0.1.0
Markdown==2.6.1
Pillow==2.8.1
South==1.0.2
app-version==0.2.1
coverage==3.7.1
django-annoying==0.8.1
django-appconf==1.0.1
django-courses==0.1.0
django-markupfield==1.3.0
django-money==0.5.0
django-object-actions==0.5.1
django-pagedown==0.1.0
django-permission==0.8.5
django-phonenumber-field==0.7.2
django-scheduler==0.7.4
eventbrite==3.0.4
future==0.14.3
icalendar==3.9.0
ipython==3.0.0
phonenumbers==7.0.4
py-moneyed==0.5.0
python-dateutil==2.4.1
pytz==2015.2
requests==2.6.0
six==1.9.0
tolerance==0.1.2
